transribe + ggfu/requirements.txt
transformers==4.44.2
python-docx==1.1.2
torchaudio==2.3.0
soundfile==0.12.1
accelerate==0.33.0
hydra-core
omegaconf
pyannote.audio
sentencepiece
llama-cpp-python==0.2.80